pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-platform"
version = "0.1.0"
description = "Agent Serving Platform — AI agent runtime at production scale"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.32,<1.0",
    "sse-starlette>=2.0,<3.0",
    # LLM providers
    "anthropic>=0.42,<1.0",
    "openai>=1.60,<2.0",
    # MCP
    "mcp>=1.0,<2.0",
    # Database
    "asyncpg>=0.30,<1.0",
    "sqlalchemy>=2.0,<3.0",
    "alembic>=1.14,<2.0",
    # Redis
    "redis[hiredis]>=5.2,<6.0",
    # Configuration
    "pydantic>=2.10,<3.0",
    "pydantic-settings>=2.7,<3.0",
    # Observability
    "opentelemetry-api>=1.29,<2.0",
    "opentelemetry-sdk>=1.29,<2.0",
    "opentelemetry-exporter-otlp>=1.29,<2.0",
    "opentelemetry-instrumentation-fastapi>=0.50b0,<1.0",
    "opentelemetry-instrumentation-httpx>=0.50b0,<1.0",
    "opentelemetry-instrumentation-redis>=0.50b0,<1.0",
    "opentelemetry-instrumentation-sqlalchemy>=0.50b0,<1.0",
    # Serialization
    "msgpack>=1.1,<2.0",
    "orjson>=3.10,<4.0",
    # Security
    "python-jose[cryptography]>=3.3,<4.0",
    "passlib[bcrypt]>=1.7,<2.0",
    "cryptography>=44,<45",
    # Utilities
    "structlog>=24.4,<25.0",
    "tenacity>=9.0,<10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9.0",
    "pytest-asyncio>=0.25,<1.0",
    "pytest-cov>=6.0,<7.0",
    "httpx>=0.28,<1.0",
    "fakeredis[lua]>=2.26,<3.0",
    "testcontainers[postgres]>=4.0,<5.0",
    "factory-boy>=3.3,<4.0",
    "ruff>=0.8,<1.0",
    "mypy>=1.14,<2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests (no external deps)",
    "integration: Integration tests (need PG + Redis)",
    "e2e: End-to-end tests (full system)",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
